third_party/blink/web_tests/fast/dom/shadow/shadowroot-keyframes-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS getBackgroundColor(shadowRootWithKeyframe.getElementById('box')) is "rgb(0, 0, 255)"
PASS getBackgroundColor(shadowRootWithoutKeyframe.getElementById('box')) is "rgb(255, 0, 0)"
